src/pdf_to_markdown_autogen/pyproject.toml
[project]
name = "pdf-to-markdown-autogen"
version = "2.0.0"
description = "AI-powered PDF to Markdown converter using AutoGen agents"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "pyautogen>=0.2.0",
    "python-dotenv>=1.0.0",
    "PyPDF2>=3.0.0",
    "pdf2image>=1.16.3",
    "Pillow>=10.0.0",
    "markdown>=3.4.0",
    "openai>=1.12.0",
]
readme = "README.md"
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["pdf_to_markdown_autogen"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3 